% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{addis_onlineFDR}
\alias{addis_onlineFDR}
\title{addis_onlineFDR}
\usage{
addis_onlineFDR(alpha, w0, raw.pvalues, gamma, lambda = 0.25, tau = 0.5)
}
\arguments{
\item{alpha}{A numeric for the desired level of type I error control.}

\item{w0}{A numeric representing the initial `wealth' of the procedure (the default values is \eqn{\tau \lambda
\alpha/2}}

\item{raw.pvalues}{A vector containing the raw p-values.}

\item{gamma}{A vector: the gamma spending sequence.}

\item{lambda}{A numeric in \link{0, 1}: the threshold for adaptivity.
Default to 0.25.}

\item{pCDFlist}{A list containing the support of the discrete CDF of the p-values.
Each support is represented by a vector in increasing order.}

\item{tau:}{A numeric in \link{0, 1}: the threshold for hypotheses to be selected for testing.
Default to 0.5.}
}
\value{
A list containing a vector of the sequence of critical values and
a vector of the indices of rejected hypothesis.
}
\description{
Function that can compute the ADDIS procedure.
This function used the function ADDIS of
the OnlineFDR package, see reference below.
For the parameters set with default values we used
the values given by Tian. J, and Ramdas. A (2019).
}
\references{
Tian, J. and Ramdas, A. (2019). ADDIS: an adaptive discarding
algorithm for online FDR control with conservative nulls.
\emph{Advances in Neural Information Processing Systems}, 9388-9396.

Robertson DS, Liou L, Ramdas A, Javanmard A, Tian J, Zrnic T, Karp NA (2019).
onlineFDR: Online error control. R package 2.1.0.
\url{https://bioconductor.org/packages/devel/bioc/html/onlineFDR.html}
}
