% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{addis_spending_onlineFDR}
\alias{addis_spending_onlineFDR}
\title{addis_spending_onlineFDR}
\usage{
addis_spending_onlineFDR(alpha, raw.pvalues, gamma, lambda = 0.25, tau = 0.5)
}
\arguments{
\item{alpha}{A numeric for the desired level of type I error control.}

\item{raw.pvalues}{A vector containing the raw p-values.}

\item{gamma}{A vector: the gamma spending sequence.}

\item{lambda}{A numeric in \link{0, 1}: the threshold for adaptivity.
Default to 0.25.}

\item{pCDFlist}{A list containing the support of the discrete CDF of the p-values.
Each support is represented by a vector in increasing order.}

\item{tau:}{A numeric in \link{0, 1}: the threshold for hypotheses to be selected for testing.
Default to 0.5.}
}
\value{
A list containing a vector of the sequence of critical values and
a vector of the indices of rejected hypothesis.
}
\description{
Function that can compute ADDIS-spending procedure.
This function used the function ADDIS_spending of
the OnlineFDR package, see reference below.
For the parameters set with default values we used
the values given by Tian. J, and Ramdas. A (2021).
}
\references{
Tian, J. and Ramdas, A. (2021). Online control of the familywise
error rate. \emph{Statistical Methods for Medical Research},
\url{https://journals.sagepub.com/eprint/AYRRKZX7XMTVHKCFYBJY/full}.

Robertson DS, Liou L, Ramdas A, Javanmard A, Tian J, Zrnic T, Karp NA (2019).
onlineFDR: Online error control. R package 2.1.0.
\url{https://bioconductor.org/packages/devel/bioc/html/onlineFDR.html}
}
