% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fwer_proc_on_impc.R
\name{compare_fwer_proc_on_impc}
\alias{compare_fwer_proc_on_impc}
\title{compare_fwer_proc_on_impc}
\usage{
compare_fwer_proc_on_impc(
  alpha,
  nb_run,
  m,
  lambda,
  gamma_type,
  q_1,
  gamma_prime_type,
  q_2
)
}
\description{
compare_fwer_proc_on_impc
}
