% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pvalues_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{data_simulation}
\usage{
data_simulation(
  N,
  m,
  non_nulls_proportion,
  p3,
  cluster_option,
  p1 = 0.01,
  p2 = 0.1,
  seed = NULL
)
}
\arguments{
\item{N}{An integer corresponding to the number of subjects studied (or the number of rows in the matrice).}

\item{m}{An integer corresponding to the number of hypotheses to test (or the number of columns in the matrice).}

\item{non_nulls_proportion}{A numeric in \link{0, 1} corresponding to the quantity of signal the user wants in the data.}

\item{p3}{A numeric in \link{0, 1} corresponding to the strength of the signal the user wants.}

\item{cluster_option}{Either "end", "begin", "begin_middle", "begin_end", "midlle_end", or "no_cluster_shuffle".
This option indicates how to position the signal in the stream of hypothesis.}

\item{p1}{A numeric corresponding to the Bernouilli parameter for generating a first group of nulls.
Default to 0.01.}

\item{p2}{A numeric corresponding to the Bernouilli parameter for generating a second group of nulls.
Default to 0.1.}

\item{seed}{An integer for setting the seed to reproduce exactly a simulation.}
}
\value{
A list containing the simulated data as a data frame and the indices of the non-nulls in the data frame.
The data frame has 4 columns of length m. The first and third columns represent
group_1 and group_2, each of their component represents a count in {0...N}.
The second and fourth columns represent the total number N which is the same for
group_1 and group_2.
}
\description{
Function that generates data for performing Fisher's two-sided exact test for marginal counts.
Detailed description on the generation of data can be found in the reference given below.
}
\examples{
data_simulation(25, 100, 0.3, 0.4, "end").

}
\references{
Sebastian Döhler, Guillermo Durand, Etienne Roquain
"New FDR bounds for discrete and heterogeneous tests,"
\emph{Electronic Journal of Statistics, Electron. J. Statist. 12(1), 1867-1900, (2018).}
\url{https://projecteuclid.org/journals/electronic-journal-of-statistics/volume-12/issue-1/New-FDR-bounds-for-discrete-and-heterogeneous-tests/10.1214/18-EJS1441.full}
}
