% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{gamma_sequence}
\alias{gamma_sequence}
\title{gamma_sequence.}
\usage{
gamma_sequence(type, nb_pvalues, q)
}
\arguments{
\item{type}{Either "log-q-serie", "q-serie" or a "rectangular" kernel.}

\item{nb_pvalues}{An integer giving the nb of p-values (/ hypothesis) to test.}

\item{q}{The exponent for computing the sequence or the kernel bandwidth.
Note that when using a rectangular kernel, q must be an integer.}
}
\value{
A vector: the gamma sequence.
}
\description{
Function that computes a nonnegative decreasing sequence.
The user can choose to make the sequence sum to exactly one
(and thus using the number of hypotheses to test),
or to make the sequence sum to less than one by approximating the infinity.
Three choices for the type of sequence are proposed, of which
log-q serie and q-serie as proposed by Tian and Ramdas (2021).
}
\references{
Tian, J. and Ramdas, A. (2021). Online control of the familywise
error rate. \emph{Statistical Methods for Medical Research},
\url{https://journals.sagepub.com/eprint/AYRRKZX7XMTVHKCFYBJY/full}
}
