% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{get_CDF}
\alias{get_CDF}
\title{get_CDF}
\usage{
get_CDF(N, m, non_nulls_proportion, p3, cluster_option, p1 = 0.01, p2 = 0.1)
}
\arguments{
\item{N}{An integer corresponding to the number of subjects studied (or the number of rows in the matrice).}

\item{m}{An integer corresponding to the number of hypotheses to test (or the number of columns in the matrice).}

\item{non_nulls_proportion}{A numeric in \link{0, 1} corresponding to the quantity of signal the user wants in the data.}

\item{p3}{A numeric in \link{0, 1} corresponding to the strength of the signal the user wants.}

\item{cluster_option}{Either "end", "begin", "begin_middle", "begin_end", "midlle_end", or "no_cluster_shuffle".
This option indicates how to position the signal in the stream of hypothesis.}

\item{p1}{A numeric corresponding to the Bernouilli parameter for generating a first group of nulls.}

\item{p2}{A numeric corresponding to the Bernouilli parameter for generating a second group of nulls.}
}
\value{
A list with the p-values' CDFs ready to use.
}
\description{
Function that allows getting the CDF of p-values ready to plot.
This function is used only for shiny apps.
}
