% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estime_power_vs_kernel_bandwidth.R
\name{kernel_bandwidth_fwer}
\alias{kernel_bandwidth_fwer}
\title{kernel_bandwidth_fwer}
\usage{
kernel_bandwidth_fwer(
  alpha,
  nb_run,
  N,
  m,
  non_nulls_proportion,
  p3,
  cluster_option,
  lambda,
  gamma_type,
  q_1,
  kernel_bandwidth,
  gamma_prime_type = "rectangular",
  p1 = 0.01,
  p2 = 0.1
)
}
\arguments{
\item{alpha}{A numeric between 0 and 1 giving the overall error budget.}

\item{nb_run}{An Integer giving the number of runs to make. If we put 100 runs, the
estimated power and fwer will be done upon a mean of 100 quantities.}

\item{N}{An integer corresponding to the number of subjects studied (or the number of rows in the matrice).}

\item{m}{An integer corresponding to the number of hypotheses to test (or the number of columns in the matrice).}

\item{non_nulls_proportion}{A numeric in \link{0, 1} corresponding to the quantity of signal the user wants in the data.}

\item{p3}{A numeric in \link{0, 1} corresponding to the strength of the signal the user wants.}

\item{cluster_option}{Either "end", "begin", "begin_middle", "begin_end", "midlle_end", or "no_cluster_shuffle".
This option indicates how to position the signal in the stream of hypothesis.}

\item{lambda}{A numeric in \link{0, 1}, to threshold the p-value for the adaptive procedures.}

\item{gamma_type}{A string, either "log-q-serie", "q-serie" or "rectangular" to specify the type of the gamma sequence.}

\item{q_1}{A numeric > 1 indicating the exponent for computing the gamma sequence.}

\item{gamma_prime_type}{A string, either "log-q-serie", "q-serie", or "rectangular" to specify the type of the gamma prime sequence.
Here the default is "rectangular" since we are studying the bandwidth for rectangular kernels.}

\item{p1}{A numeric corresponding to the Bernouilli parameter for generating a first group of nulls.
Defaults to 0.01.}

\item{p2}{A numeric corresponding to the Bernouilli parameter for generating a second group of nulls.
Default to 0.1.}

\item{q_2}{An integer > 1 giving the bandwidth of the rectangular kernel.}

\item{tau}{A numeric in \link{0, 1}, used as a  threshold for hypotheses to be selected for testing.
Used for Addis-spending only, kept Tian, J. and Ramdas, A. (2021) default values 0.5.}
}
\value{
Returns two data frames.
- The first data frame contains the ratio between the number of true discoveries and the number of non-nulls
for each run performed for each procedure.
This ratio must be averaged over all runs, for each procedure to estimate the power.
- The second data frame contains an estimate of the mFDR for each procedure i.e. the ratio between
the mean (over nb_run) of the number of false discoveries
and the mean (over nb_run) of the number of discoveries.
}
\description{
Function used to compare the performance of rhoOB and rhoAOB for
different rectangular kernel bandwidths.
}
