% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lord_OnlineSuperUnif}
\alias{lord_OnlineSuperUnif}
\title{lord_OnlineSupeUnif}
\usage{
lord_OnlineSuperUnif(alpha, w0, raw_pvalues, gamma)
}
\arguments{
\item{alpha}{A numeric for the desired level of type I error control.}

\item{w0}{A numeric in \link{0, alpha} representing the amount of earning back one wants.}

\item{gamma}{A vector: the gamma spending sequence.}

\item{raw.pvalues}{A vector containing raw pvalues.}

\item{pCDFlist}{A list containing the support of the discrete CDF of the p-values.
Each support is represented by a vector in increasing order.}
}
\value{
A list containing a vector of the sequence of critical values and
a vector of the indices of rejected hypothesis.
}
\description{
Function that computes LORD of Ramdas et al. (2017).
}
\references{
Ramdas, A., Yang, F., Wainwright M.J. and Jordan, M.I. (2017). Online control
of the false discovery rate with decaying memory. \emph{Advances in Neural
Information Processing Systems 30}, 5650-5659.
}
