% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OB_AOB_Rcpp}
\alias{OB_AOB_Rcpp}
\title{OB_AOB_Rcpp}
\usage{
OB_AOB_Rcpp(alpha, raw_pvalues, gamma, lambda)
}
\arguments{
\item{alpha}{A numeric in \link{0, 1} for the desired level of type I error control.}

\item{gamma}{A vector : the gamma spending sequence.}

\item{lambda}{A numeric in \link{0, 1} : the adaptivity threshold that sets the
threshold for 'candidate' hypothesis. Default to 0.5,
setting lambda to 0 gives 0B the non adaptive versio of AOB.}

\item{raw.pvalues}{A vector containing raw p-values.}
}
\description{
Function that can compute OB and AOB (Adaptive-Spending of Tian & Ramdas (2021)) procedure.
Setting lambda to 0 gives OB.
}
\references{
Tian, J. and Ramdas, A. (2021). Online control of the familywise
error rate. \emph{Statistical Methods for Medical Research},
\url{https://journals.sagepub.com/eprint/AYRRKZX7XMTVHKCFYBJY/full}.
}
