% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pvalues_simulation.R
\name{pvalues_simulation}
\alias{pvalues_simulation}
\title{pvalues-simulation}
\usage{
pvalues_simulation(dataframe, alternative = "two.sided", input = "marginal")
}
\arguments{
\item{alternative}{A string informing if we are doing two.sided test or not.
The argument can either be "two.sided", "less", or "greater".
Here the default is to "two.sided".}

\item{input}{The format of the input data frame, see the references below.
Here the default is to "marginal".}

\item{data}{frame A data frame of 4 columns and any number of rows (the number of rows
is the number of hypotheses that the user is testing).}
}
\value{
A list containing the raw p-values and their CDF support.
}
\description{
Function that performs Fisher's exact two-sided test.
This function uses the 'fisher.pvalues.support' from
the DiscreteFDR package of Junge. F et al (2019).
}
\examples{
test_data <- data.frame(c(1, 0, 2, 5, 2), rep(25, 5), c(0, 0, 2, 14, 11), rep(25, 5))
          or test_data <- data_simulation(25, 100, 0.3, 0.4, "end")$data
          pvalues_simulation(test_data).

}
\references{
Junge. F, Durand. G, Döhler. S, and Roquain. E (2019)
DiscreteFDR: An R package for controlling the false
discovery rate for discrete test statistics.
\url{https://arxiv.org/pdf/1904.02054.pdf}
\url{https://cran.r-project.org/web/packages/DiscreteFDR/index.html}.
}
