% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{rho_ALORD}
\alias{rho_ALORD}
\title{rho_ALORD}
\usage{
rho_ALORD(alpha, w0, raw.pvalues, pCDFlist, gamma, lambda, gamma_prime)
}
\arguments{
\item{alpha}{A numeric in \link{0, 1} for the desired level of type I error control.}

\item{w0}{A numeric representing the initial wealth.}

\item{raw.pvalues}{A vector containing the raw p-values.}

\item{pCDFlist}{A list containing the support of the discrete CDF of the p-values.
Each support is represented by a vector in increasing order.}

\item{gamma}{A vector: the gamma spending sequence.}

\item{gamma_prime}{A vector: the gamma prime smoothing sequence (it can be the same as gamma).
When gamma_prime is not provided, the greedy version of the procedure is performed.}
}
\value{
A list containing a vector of the sequence of critical values and
a vector of the indices of rejected hypothesis.
}
\description{
Function that can compute rho-ALORD procedure.
}
