% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_OB_AOB_Rcpp}
\alias{rho_OB_AOB_Rcpp}
\title{rho_OB_AOB_Rcpp}
\usage{
rho_OB_AOB_Rcpp(
  alpha,
  raw_pvalues,
  pCDFlist,
  gamma,
  lambda,
  gamma_prime = numeric(),
  greedy = TRUE
)
}
\arguments{
\item{alpha}{A numeric for the desired level of type I error control.}

\item{pCDFlist}{A list containing the support of the discrete CDF of the p-values.
Each support is represented by a vector in increasing order.}

\item{gamma}{A vector: the gamma spending sequence.}

\item{lambda}{A numeric in \link{0, 1} : the threshold for adaptivity.}

\item{gamma_prime}{A vector: the gamma prime smoothing sequence (it can be the same as gamma).}

\item{greedy}{A boolean to indicate whether we want to perform the greedy version (TRUE) or not (FALSE).
Default to TRUE.}

\item{raw.pvalues}{A vector containing raw pvalues.}
}
\value{
A list containing a vector of the sequence of critical values and
a vector of the indices of rejected hypothesis.
}
\description{
Function that can compute rho-OB and rho-AOB.
To compute rho_OB, lambda must be 0 (i.e. we don't threshold for adaptivity to the number of true nulls)
When gamma_prime is not provided and 'greedy' set to true, the greedy version of the procedure is computed.
}
