% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{shuffle_vec}
\alias{shuffle_vec}
\title{shuffle_vec}
\usage{
shuffle_vec(vec, permutation_index = NULL)
}
\arguments{
\item{vec}{A vector that needs to be shuffled.}

\item{permutation_index}{A vector indicating how to shuffle the vector
if one wants to perform a certain permutation.}
}
\value{
A list containing the shuffled vector and the index of the entries.
}
\description{
Function that shuffles a vector (permutation).
This function allows to study the signal position scheme where
the signal is not clustered but positioned randomly across the whole stream of hypothesis
(signal position = "no_cluster_shuffle" in data_simulation function).
}
