% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stepfun}
\alias{stepfun}
\title{stepfun}
\usage{
stepfun(x, sfun)
}
\description{
Fast step function evaluation.
Step function is represented by a single numeric vector under the conditions
a) f(x) = x and b) it is complete, meaning that the last value is 1,
this is much faster than passing and evaluating R step function objects.
}
